Create meetingrooms3.cpp
Q. Given an integer n representing the number of rooms numbered from 0 to n - 1 and a 2D integer array meetings[][] where meetings[i] = [starti, endi] indicates that a meeting is scheduled during the half-closed time interval [starti, endi]. The task is to find the room number that hosts the most meetings. If multiple rooms have the same highest number of meetings, find the smallest room number among them.

Meeting Allocation Rules

When a meeting starts, assign it to the available room with the smallest number.
If no rooms are free, delay the meeting until the earliest room becomes available. The delayed meeting retains its original duration.
When a room becomes free, assign it to the delayed meeting with the earliest original start timing.


Approach-Using Two priority queue 
 O(m*log m+m*log n) Time 
 O(n) Space